+++ src/Agents.jl
@@ -27,6 +27,7 @@ include("simulations/sample.jl")
 
 # 4.0 Depreciations
 @deprecate space_neighbors nearby_agents
+@deprecate node_neighbors nearby_positions
 
 # Predefined models
 include("models/Models.jl")